Rename tgo package to gocode
