+++ encoding/gocode/binding.tmpl
